requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
anthropic>=0.3.0
python-dotenv>=1.0.0
pandas>=2.0.0
openpyxl>=3.1.0
